Renaming components & tags utils to TS
