+++ src/lib/components/MaintenanceStatusIcon.js
@@ -1,5 +1,5 @@
 import { BaseComponent } from "$lib/components/base/BaseComponent";
-import { getComponent } from "$lib/components/base/ComponentUtils";
+import { getComponent } from "$lib/components/base/component-utils";
 import { on } from "$lib/components/events/comms";
 import { eventMaintenanceStateChanged } from "$lib/components/events/maintenance-popup-events";
